fix: unify link edit modals
